Shared Kernelを導入し、監査ログや在庫、注文に関するドメインモデルを整理しました。AuditLogエンティティを削除し、AuditLogRecordを使用するように変更。InventoryとOrderエンティティにValue Objectを導入し、ドメインロジックを強化しました。また、リポジトリの実装を改善し、マッピング処理を追加しました。

+++ src/Web.Api/Contracts/Inventories/Responses/InventoryMappingExtensions.cs
@@ -14,7 +14,7 @@ public static class InventoryMappingExtensions
     /// <returns>在庫情報のレスポンスDTO</returns>
     public static InventoryResponse ToResponse(this Inventory inventory) =>
         new(
-            inventory.ProductId,
+            (int)inventory.ProductId,
             inventory.ProductName,
             inventory.Stock,
             inventory.UnitPrice);
